project_requirements.txt
email-validator>=2.2.0
flask-wtf>=1.2.2
flask>=3.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
werkzeug>=3.1.3
flask-login>=0.6.3
wtforms>=3.2.1
openai>=1.78.0
markupsafe>=3.0.2
sqlalchemy>=2.0.40
trafilatura>=2.0.0
flask-migrate>=4.1.0
anthropic>=0.51.0
sendgrid>=6.12.0
twilio>=9.6.0
stripe>=12.1.0